+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 #### 4.63.3 - 12.12.2017
+* New helper: GetDotNetSDKVersionFromGlobalJson
 * Do not download DotNet SDK when it's already in temp
 
 #### 4.63.2 - 11.09.2017